requirements.txt
openai
google-genai
qdrant-client
fastembed
langchain
sentence-transformers
fastapi
datasets
python-dotenv
uvicorn